move functions for execution report to report/execution_report.py
