+++ golem/report/report.py
@@ -1,16 +1,10 @@
 import base64
 import json
 import os
-import sys
 import shutil
-import traceback
 
 from golem.core import session, utils
-from golem.test_runner.conf import ResultsEnum
 from golem.core.project import Project
-
-## DELETE
-from golem.gui.report_parser import get_date_time_from_timestamp, get_execution_report_default
 
 
 def get_last_executions(projects=None, suite=None, limit=5):
@@ -48,120 +42,6 @@ def get_last_executions(projects=None, suite=None, limit=5):
             if len(last_executions):
                 last_execution_data[project][s] = last_executions
     return last_execution_data
-
-
-def _parse_execution_data(execution_directory=None, project=None,
-                          suite=None, execution=None, finalize=False):
-    execution_data = get_execution_report_default()
-    if not execution_directory:
-        execution_directory = os.path.join(session.testdir, 'projects', project,
-                                           'reports', suite, execution)
-    tests = []
-    if os.path.isdir(execution_directory):
-        tests = next(os.walk(execution_directory))[1]
-
-    for test in tests:
-        # each test may have n >= 1 test sets
-        # each test set is considered a different test
-        test_path = os.path.join(execution_directory, test)
-        test_sets = os.walk(test_path).__next__()[1]
-
-        for test_set in test_sets:
-            new_test = {
-                'test_set': '',
-                'module': '',
-                'sub_modules': '',
-                'name': '',
-                'full_name': '',
-                'result': '',
-                'test_elapsed_time': '',
-                'start_date_time': '',
-                'browser': '',
-                'data': '',
-                'environment': '',
-                'set_name': ''
-            }
-            execution_data['total_tests'] += 1
-            new_test['test_set'] = test_set
-            module = ''
-            sub_modules = []
-            test_split = test.split('.')
-            if len(test_split) > 1:
-                module = test_split[0]
-                if len(test_split) > 2:
-                    sub_modules = test_split[1:-1]
-            new_test['module'] = module
-            test_name = test_split[-1]
-            new_test['sub_modules'] = sub_modules
-            new_test['name'] = test_name
-            new_test['full_name'] = test
-
-            test_set_path = os.path.join(test_path, test_set)
-            report_json_path = os.path.join(test_set_path, 'report.json')
-            report_log_path = os.path.join(test_set_path, 'execution_info.log')
-            try:
-                with open(report_json_path) as f:
-                    report_data = json.load(f)
-                    status = report_data['result']
-                    new_test['test_elapsed_time'] = report_data['test_elapsed_time']
-                    start_date_time = get_date_time_from_timestamp(
-                        report_data['test_timestamp'])
-                    new_test['start_date_time'] = start_date_time
-                    new_test['browser'] = report_data['browser']
-                    new_test['data'] = report_data['test_data']
-                    new_test['environment'] = report_data['environment']
-                    # TODO, previous versions won't have set_name
-                    # remove the if when retro-compatibility is not required
-                    if 'set_name' in report_data:
-                        new_test['set_name'] = report_data['set_name']
-            except FileNotFoundError:
-                if os.path.isfile(report_log_path):
-                    # test had been started
-                    status = ResultsEnum.STOPPED if finalize else ResultsEnum.RUNNING
-                else:
-                    # test had not been started
-                    status = ResultsEnum.NOT_RUN if finalize else ResultsEnum.PENDING
-            except json.decoder.JSONDecodeError:
-                # report.json exists but contains malformed JSON
-                status = ResultsEnum.STOPPED if finalize else ResultsEnum.RUNNING
-            except Exception:
-                sys.exit('an error occurred generating JSON report\n{}'
-                         .format(traceback.format_exc()))
-            new_test['result'] = status
-            _status_total = execution_data['totals_by_result'].get(status, 0) + 1
-            execution_data['totals_by_result'][status] = _status_total
-            execution_data['tests'].append(new_test)
-    return execution_data
-
-
-def get_execution_data(execution_directory=None, project=None, suite=None,
-                       execution=None):
-    """Retrieve the data of all the tests of a suite execution.
-
-    From the report.json if it exists, otherwise it parses
-    the tests one by one.
-
-    The `report.json` should be generated when the suite
-    execution ends.
-    """
-    has_finished = False
-    if execution_directory is None:
-        execution_directory = os.path.join(session.testdir, 'projects', project,
-                                           'reports', suite, execution)
-    if os.path.isfile(os.path.join(execution_directory, 'report.json')):
-        with open(os.path.join(execution_directory, 'report.json')) as f:
-            data = json.load(f)
-            has_finished = True
-    elif os.path.isfile(os.path.join(execution_directory, 'execution_report.json')):
-        # backward compatibility
-        # TODO remove
-        with open(os.path.join(execution_directory, 'execution_report.json')) as f:
-            data = json.load(f)
-            has_finished = True
-    else:
-        data = _parse_execution_data(execution_directory, project, suite, execution)
-    data['has_finished'] = has_finished
-    return data
 
 
 def get_test_case_data(project, test, suite=None, execution=None, test_set=None,
@@ -217,7 +97,7 @@ def get_test_case_data(project, test, suite=None, execution=None, test_set=None,
             test_data['description'] = report_data['description']
             test_data['result'] = report_data['result']
             test_data['test_elapsed_time'] = report_data['test_elapsed_time']
-            start_date_time = get_date_time_from_timestamp(report_data['test_timestamp'])
+            start_date_time = utils.get_date_time_from_timestamp(report_data['test_timestamp'])
             test_data['start_date_time'] = start_date_time
             test_data['errors'] = report_data['errors']
             test_data['browser'] = report_data['browser']
